// src/components/post/__generated__/PostView_post_Fragment.graphql.ts
/**
 * @generated SignedSource<<3e09cb4fee8ef112c9461f49343fac5e>>
 * @lightSyntaxTransform
 * @nogrep
 */

/* tslint:disable */
/* eslint-disable */
// @ts-nocheck

import { Fragment, ReaderFragment } from 'relay-runtime';
export type ContentType = "Markdown" | "PlainText" | "%future added value";
import { FragmentRefs } from "relay-runtime";
export type PostView_post_Fragment$data = {
  readonly id: string;
  readonly postedAt: any;
  readonly title: string;
  readonly type: ContentType;
  readonly category: string;
  readonly content: string;
  readonly author: {
    readonly id: string;
    readonly name: string;
    readonly avatar: string;
  } | null;
  readonly " $fragmentType": "PostView_post_Fragment";
};
export type PostView_post_Fragment$key = {
  readonly " $data"?: PostView_post_Fragment$data;
  readonly " $fragmentSpreads": FragmentRefs<"PostView_post_Fragment">;
};

const node: ReaderFragment = (function(){
var v0 = {
  "alias": null,
  "args": null,
  "kind": "ScalarField",
  "name": "id",
  "storageKey": null
};
return {
  "argumentDefinitions": [],
  "kind": "Fragment",
  "metadata": null,
  "name": "PostView_post_Fragment",
  "selections": [
    (v0/*: any*/),
    {
      "alias": null,
      "args": null,
      "kind": "ScalarField",
      "name": "postedAt",
      "storageKey": null
    },
    {
      "alias": null,
      "args": null,
      "kind": "ScalarField",
      "name": "title",
      "storageKey": null
    },
    {
      "alias": null,
      "args": null,
      "kind": "ScalarField",
      "name": "type",
      "storageKey": null
    },
    {
      "alias": null,
      "args": null,
      "kind": "ScalarField",
      "name": "category",
      "storageKey": null
    },
    {
      "alias": null,
      "args": null,
      "kind": "ScalarField",
      "name": "content",
      "storageKey": null
    },
    {
      "alias": null,
      "args": null,
      "concreteType": "User",
      "kind": "LinkedField",
      "name": "author",
      "plural": false,
      "selections": [
        (v0/*: any*/),
        {
          "alias": null,
          "args": null,
          "kind": "ScalarField",
          "name": "name",
          "storageKey": null
        },
        {
          "alias": null,
          "args": null,
          "kind": "ScalarField",
          "name": "avatar",
          "storageKey": null
        }
      ],
      "storageKey": null
    }
  ],
  "type": "Post",
  "abstractKey": null
};
})();

(node as any).hash = "d507424cec259bb0a970c3a1039f464b";

export default node;